\name{calcClosedChamberFluxForChunks}
\alias{calcClosedChamberFluxForChunks}
\title{calcClosedChamberFluxForChunks}
\description{apply \code{\link{calcClosedChamberFlux}} for each chunk in data}
\usage{calcClosedChamberFluxForChunks(ds, colChunk = "iChunk", ...)}\arguments{
  \item{ds}{tibble or data.frame}
  \item{colChunk}{column name (scalar string) holding a factor unique to each chunk}
  \item{\dots}{
}
}



\author{Thomas Wutzler, Oscar Perez Priego}





