\name{regressSelectAIC}
\alias{regressSelectAIC}
\title{regressSelectAIC}
\description{select regression function based on minimum AIC of the fits	}
\usage{regressSelectAIC(fluxEstL)}
\arguments{
  \item{fluxEstL}{list of return values of different regression functions such as \code{\link{regressFluxLinear}}}
}

\value{index of the best regression function}

\author{Thomas Wutzler, Oscar Perez Priego}





