\name{regressSelectPref1}
\alias{regressSelectPref1}
\title{regressSelectPref1}
\description{prefer first regression function, if its AIC is not significantly different from next best 	}
\usage{regressSelectPref1(fluxEstL)}
\arguments{
  \item{fluxEstL}{list of return values of different regression functions such as \code{\link{regressFluxLinear}}}
}

\value{index of the best regression function}

\author{Thomas Wutzler, Oscar Perez Priego}





